prepares app for google play store release
