+++ app/build.gradle
@@ -3,7 +3,7 @@ apply plugin: 'com.android.application'
 android {
     compileSdkVersion 27
     defaultConfig {
-        applicationId "com.example.andresprato.venezueladream"
+        applicationId "com.mobileapp.andresprato.venezueladream"
         minSdkVersion 21
         targetSdkVersion 27
         versionCode 1
