--- app/src/main/java/com/example/andresprato/venezueladream/PaymentInformationActivity.java
@@ -1,58 +1,0 @@
-package com.example.andresprato.venezueladream;
-
-import android.content.DialogInterface;
-import android.os.Bundle;
-import android.support.design.widget.FloatingActionButton;
-import android.support.v7.app.AlertDialog;
-import android.support.v7.app.AppCompatActivity;
-import android.view.View;
-
-import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.auth.FirebaseUser;
-
-public class PaymentInformationActivity extends AppCompatActivity{
-    private FloatingActionButton fab;
-
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_payment_information);
-
-        getSupportActionBar().setDisplayHomeAsUpEnabled(true);
-        getSupportActionBar().setDisplayShowHomeEnabled(true);
-
-        FirebaseUser user = FirebaseAuth.getInstance().getCurrentUser();
-
-        fab = findViewById(R.id.fab);
-
-        fab.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                //user.updateProfile();
-
-                alert();
-
-
-            }
-        });
-    }
-
-    public void alert(){
-        AlertDialog.Builder builder = new AlertDialog.Builder(this);
-        builder.setMessage("You just saved your credit card details!")
-                .setTitle("Save Credit Card Details")
-                .setPositiveButton("Close", new DialogInterface.OnClickListener() {
-                    public void onClick(DialogInterface dialog, int id) {
-                        finish();
-                    }
-                });
-
-        AlertDialog dialog = builder.create();
-        dialog.show();
-    }
-
-    @Override
-    public boolean onSupportNavigateUp() {
-        finish();
-        return true;
-    }
-}
