--- app/src/main/java/com/example/andresprato/venezueladream/WelcomeActivity.java
@@ -1,4 +1,0 @@
-package com.example.andresprato.venezueladream;
-
-public class WelcomeActivity {
-}
