+++ app/src/main/java/com/mobileapp/andresprato/venezueladream/WelcomeActivity.java
@@ -1,0 +1,4 @@
+package com.mobileapp.andresprato.venezueladream;
+
+public class WelcomeActivity {
+}
