+++ app/src/main/res/xml/pref_headers.xml
@@ -3,17 +3,17 @@
     <!-- These settings headers are only used on tablets. -->
 
     <header
-        android:fragment="com.example.andresprato.venezueladream.SettingsAct$GeneralPreferenceFragment"
+        android:fragment="com.mobileapp.andresprato.venezueladream.SettingsAct$GeneralPreferenceFragment"
         android:icon="@drawable/ic_info_black_24dp"
         android:title="@string/pref_header_general" />
 
     <header
-        android:fragment="com.example.andresprato.venezueladream.SettingsAct$NotificationPreferenceFragment"
+        android:fragment="com.mobileapp.andresprato.venezueladream.SettingsAct$NotificationPreferenceFragment"
         android:icon="@drawable/ic_notifications_black_24dp"
         android:title="@string/pref_header_notifications" />
 
     <header
-        android:fragment="com.example.andresprato.venezueladream.SettingsAct$DataSyncPreferenceFragment"
+        android:fragment="com.mobileapp.andresprato.venezueladream.SettingsAct$DataSyncPreferenceFragment"
         android:icon="@drawable/ic_sync_black_24dp"
         android:title="@string/pref_header_data_sync" />
